Add iptables rule to avoid conntrack for proxy's forward packets
To ensure IPsec encryption for proxy forward packets, we added routing
rule to push them to cilium_host. This change caused side effects for
to-world traffic. This patch fixes the consequences of side effects.

Proxy will perform SNAT for to-world packets, but the new source address
is decided by routing rules. Previously, to-world packets are routed to
eth0, so proxy uses eth0's address for SNAT. Now with new routing rule
to push them to cilium_host instead of eth0, proxy uses cilium_host's
address for SNAT as the side effect.

This change makes to-world packets rely on "external" SNAT, which wasn't
required because proxy's SNAT worked perfectly. We need "external" SNAT
to change source address of to-world packets from cilium_host's IP to
eth0's IP. As IPsec doesn't work with KPR, the "external" SNAT mechanism
is iptables.

However, due to kernel's implementation details, an skb won't be
processed by nat POSTROUTING for twice. When the packet is routed to
cilium_host, it's the first time; when forwarded from cilium_net to
eth0, it's supposed to be the second time.

This is because, After the first POSTROUTING traversal, skb's ct (struct
nf_conn*)(skb->_nfct & ~7) has a status IPS_SRC_NAT_DONE to skip the
second traversal at all.

To avoid setting the IPS_SRC_NAT_DONE flag, this patch adds an iptables
rule `-j CT --notrack`.

Signed-off-by: Zhichuan Liang <gray.liang@isovalent.com>