ci: use base and head SHAs from context in lint-build-commits workflow
Instead of querying the GitHub API for the parent SHA, use the base and
head SHA provided by the github.event.pull_request context. This works
fine because the workflow only runs on pull requests.

Also use a loop and git checkout instead of git rebase to avoid
potential issues with merge conflicts in the presence of merge commits
in the PR.

Signed-off-by: Tobias Klauser <tobias@cilium.io>